Adding support for screens.  This will let us select a mission from a list
